extract/radar to Radar.vue
